Factor out str-template module so that it can be used from compiler and engine
