+++ modules/engine/lib/engine/source/verb.js
@@ -22,7 +22,7 @@ var assert = require('assert'),
     headers = require('headers'),
     async = require('async'),
     MutableURI = require('ql.io-mutable-uri'),
-    strTemplate = require('../peg/str-template.js'),
+    strTemplate = require('ql.io-str-template'),
     uriTemplate = require('ql.io-uri-template'),
     fs = require('fs'),
     normalize = require('path').normalize,
