+++ modules/str-template/.npmignore
@@ -1,0 +1,2 @@
+node_modules
+test
