+++ modules/str-template/lib/README.md
