+++ modules/str-template/lib/str-parser.js
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2011 eBay Software Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+var strTemplate = require('./str-template.js');
+
+'use strict'
+
+var cache = {};
+exports.parse = function(template) {
+    var parsed = cache[template];
+    if(parsed) {
+        return parsed;
+    }
+
+    parsed = strTemplate.parse(template);
+    cache[template] = parsed;
+    return parsed;
+}
