+++ modules/str-template/package.json
@@ -1,20 +1,20 @@
 {
     "author": "ql.io",
-    "name": "ql.io-compiler",
-    "version": "0.6.1",
+    "description": "A small string template processor for ql.io.",
+    "name": "ql.io-str-template",
+    "version": "0.6.0",
     "repository": {
         "type": "git",
         "url": "https://github.com/ql-io/ql.io"
     },
+    "main": "lib/str-parser.js",
     "engines": {
         "node":"> 0.6.0"
     },
-    "main": "lib/compiler.js",
     "dependencies": {
-        "ql.io-str-template": "",
-        "underscore": "1.3.3"
     },
     "devDependencies": {
-        "nodeunit": "0.7.4"
+        "nodeunit": "0.7.4",
+        "underscore": "1.3.3"
     }
 }
